removed duplicate code and tests
