+++ src/main/java/model/Group.java
@@ -6,6 +6,8 @@ import java.util.HashMap;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
+import model.datafield.DataField;
+
 /**
  * This class represents a group of files.
  * @author Hans Schouten
