+++ src/test/java/analyze/comparing/ComparerTest.java
@@ -7,15 +7,15 @@ import java.text.ParseException;
 
 import model.Column;
 import model.ColumnType;
-import model.DataField;
-import model.DataFieldDouble;
-import model.DataFieldString;
 import model.DateColumn;
 import model.DateUtils;
 import model.Reader;
 import model.Record;
 import model.RecordList;
 import model.SequentialData;
+import model.datafield.DataField;
+import model.datafield.DataFieldDouble;
+import model.datafield.DataFieldString;
 
 import org.junit.Before;
 import org.junit.Test;
@@ -71,7 +71,8 @@ public class ComparerTest {
                 
         actual = comparer.compare();
 
-        assertEquals(expected.last().get("Time difference").getStringValue(), actual.last().get("Time difference").getStringValue());
+        assertEquals(expected.last().get("Time difference").toString(),
+                actual.last().get("Time difference").toString());
 
     }
     
@@ -83,7 +84,8 @@ public class ComparerTest {
     public void testParseTimeBetween() throws AnalyzeException, ParseException {
         result = p.parse("COMPARE datum1 AND datum2", userData);
             
-        assertEquals(expected.last().get("Time difference").getStringValue(), result.last().get("Time difference").getStringValue());
+        assertEquals(expected.last().get("Time difference").toString(), 
+                result.last().get("Time difference").toString());
 
     }
     
@@ -97,7 +99,7 @@ public class ComparerTest {
         
         String expected = "-0y0m-5d -3h-5m";
         
-        assertEquals(expected, result.last().get("Time difference").getStringValue());
+        assertEquals(expected, result.last().get("Time difference").toString());
 
     }
