+++ src/test/java/analyze/condition/BinaryOpTermTest.java
@@ -2,11 +2,11 @@ package analyze.condition;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
-import model.DataFieldBoolean;
-import model.DataFieldDouble;
-import model.DataFieldInt;
-import model.DataFieldString;
 import model.UnsupportedFormatException;
+import model.datafield.DataFieldBoolean;
+import model.datafield.DataFieldDouble;
+import model.datafield.DataFieldInt;
+import model.datafield.DataFieldString;
 
 import org.junit.Test;
