+++ src/test/java/analyze/parsing/ConstrainParserTest.java
@@ -1,10 +1,10 @@
 package analyze.parsing;
 
 import static org.junit.Assert.*;
-import model.DataFieldDouble;
 import model.DateUtils;
 import model.Record;
 import model.SequentialData;
+import model.datafield.DataFieldDouble;
 
 import org.junit.Before;
 import org.junit.Test;
